backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-ninja==1.3.0
pydantic==2.9.1
pydantic_core==2.23.3
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.43.0
